#428 Move Calendar to Program component, show all day acitivities correctly in agenda view, remove Acitivity component
