+++ webapp/src/components/CalendarPage.tsx
@@ -1,17 +1,9 @@
-import {
-  Calendar,
-  luxonLocalizer,
-  Event as RBCEvent,
-} from 'react-big-calendar';
-import { Container } from '@mui/material';
-import { DateTime, Settings } from 'luxon';
-import { ProgramWithActivities } from '../types/api';
 import React from 'react';
-import 'react-big-calendar/lib/css/react-big-calendar.css';
-import useApiData from '../helpers/useApiData';
+import { Container } from '@mui/material';
 
-Settings.defaultZone = 'America/Los_Angeles';
-const localizer = luxonLocalizer(DateTime);
+import Program from './Program';
+import useApiData from '../helpers/useApiData';
+import { ProgramWithActivities } from '../types/api';
 
 const CalendarPage = () => {
   const {
@@ -24,33 +16,36 @@ const CalendarPage = () => {
     sendCredentials: true,
   });
   if (error) {
-    return <p>There was an error loading the programs.</p>;
+    return (
+      <Container fixed>
+        <p>There was an error loading the programs.</p>
+      </Container>
+    );
   }
   if (isLoading) {
-    return <div>Loading...</div>;
+    return (
+      <Container fixed>
+        <div>Loading...</div>
+      </Container>
+    );
   }
   if (!programs) {
     return null;
   }
-  const programEventsActivities: RBCEvent[] = programs[0].activities.map(
-    activity => {
-      return {
-        title: activity.title,
-        start: new Date(activity.start_time),
-        end: new Date(activity.end_time),
-        description: activity.description_text,
-      };
-    }
-  );
   return (
     <Container fixed>
-      <Calendar
-        localizer={localizer}
-        events={programEventsActivities}
-        startAccessor="start"
-        endAccessor="end"
-        style={{ height: 500 }}
-      />
+      <h2>Programs</h2>
+      {programs.map(program => {
+        return (
+          <Program
+            key={program.id}
+            title={program.title}
+            startDate={program.start_date}
+            endDate={program.end_date}
+            activities={program.activities}
+          />
+        );
+      })}
     </Container>
   );
 };
